Reorganize how records are submitted to collections in frontend (#141 #142)
This restores the ability to add records to collections and also
improves the separation of concerns. The VRECollectionView is
responsible for selecting collections and for detecting when the user
wants to submit. Its parent views are responsible for passing the URIs
of the models that should be added to those collections.