Temporarily downgrade sync method in frontend Annotation models (#254 #287)
The backend endpoint has a parser override that downgrades from
application/ld+json to application/json. With the frontend now sending
proper JSON-LD, this override has become counterproductive. The
correct fix would be to remove the override in the backend, but that
requires more code changes. This should (temporarily) address the
problem reported by @tijmenbaarda in
https://github.com/CentreForDigitalHumanities/EDPOP/pull/287#pullrequestreview-3257120667.